+++ Classes/Model/Type/Accommodation.php
@@ -16,8 +16,6 @@ use Brotkrueml\Schema\Core\Model\AbstractType;
 
 /**
  * An accommodation is a place that can accommodate human beings, e.g. a hotel room, a camping pitch, or a meeting room. Many accommodations are for overnight stays, but this is not a mandatory requirement.
- * For more specific types of accommodations not defined in schema.org, one can use additionalType with external vocabularies.
- *
  * See also the <a href="https://schema.org/docs/hotels.html">dedicated document on the use of schema.org for marking up hotels and other forms of accommodations</a>.
  */
 #[Type('Accommodation')]