+++ Classes/Model/Type/GeoCircle.php
@@ -16,7 +16,6 @@ use Brotkrueml\Schema\Core\Model\AbstractType;
 
 /**
  * A GeoCircle is a GeoShape representing a circular geographic area. As it is a GeoShape
- * it provides the simple textual property 'circle', but also allows the combination of postalCode alongside geoRadius.
  * The center of the circle can be indicated via the 'geoMidpoint' property, or more approximately using 'address', 'postalCode'.
  */
 #[Type('GeoCircle')]