+++ Classes/Model/Type/Place.php
@@ -11,13 +11,16 @@ declare(strict_types=1);
 
 namespace Brotkrueml\Schema\Model\Type;
 
+use Brotkrueml\Schema\Attributes\Manual;
 use Brotkrueml\Schema\Attributes\Type;
 use Brotkrueml\Schema\Core\Model\AbstractType;
+use Brotkrueml\Schema\Manual\Publisher;
 
 /**
  * Entities that have a somewhat fixed, physical extension.
  */
 #[Type('Place')]
+#[Manual(Publisher::Yandex, 'https://yandex.com/support/webmaster/supported-schemas/address-organization.html')]
 final class Place extends AbstractType
 {
     protected static array $propertyNames = [